refactor(agent): extract turn loop into agent/turn.rs
